extracted the type checker into a separate module
