+++ src/isomorphic/classic/element/ReactElementValidator.js
@@ -21,12 +21,12 @@
 var ReactCurrentOwner = require('ReactCurrentOwner');
 var ReactComponentTreeDevtool = require('ReactComponentTreeDevtool');
 var ReactElement = require('ReactElement');
-var ReactPropTypeLocationNames = require('ReactPropTypeLocationNames');
 var ReactPropTypeLocations = require('ReactPropTypeLocations');
+
+var checkTypes = require('checkTypes');
 
 var canDefineProperty = require('canDefineProperty');
 var getIteratorFn = require('getIteratorFn');
-var invariant = require('invariant');
 var warning = require('warning');
 
 function getDeclarationErrorAddendum() {
@@ -45,8 +45,6 @@ function getDeclarationErrorAddendum() {
  * updates.
  */
 var ownerHasKeyUseWarning = {};
-
-var loggedTypeFailures = {};
 
 function getCurrentComponentErrorInfo(parentType) {
   var info = getDeclarationErrorAddendum();
@@ -153,66 +151,6 @@ function validateChildKeys(node, parentType) {
 }
 
 /**
- * Assert that the props are valid
- *
- * @param {object} element
- * @param {string} componentName Name of the component for error messages.
- * @param {object} propTypes Map of prop name to a ReactPropType
- * @param {string} location e.g. "prop", "context", "child context"
- * @private
- */
-function checkPropTypes(element, componentName, propTypes, location) {
-  var props = element.props;
-  for (var propName in propTypes) {
-    if (propTypes.hasOwnProperty(propName)) {
-      var error;
-      // Prop type validation may throw. In case they do, we don't want to
-      // fail the render phase where it didn't fail before. So we log it.
-      // After these have been cleaned up, we'll let them throw.
-      try {
-        // This is intentionally an invariant that gets caught. It's the same
-        // behavior as without this statement except with a better message.
-        invariant(
-          typeof propTypes[propName] === 'function',
-          '%s: %s type `%s` is invalid; it must be a function, usually from ' +
-          'React.PropTypes.',
-          componentName || 'React class',
-          ReactPropTypeLocationNames[location],
-          propName
-        );
-        error = propTypes[propName](props, propName, componentName, location);
-      } catch (ex) {
-        error = ex;
-      }
-      warning(
-        !error || error instanceof Error,
-        '%s: type specification of %s `%s` is invalid; the type checker ' +
-        'function must return `null` or an `Error` but returned a %s. ' +
-        'You may have forgotten to pass an argument to the type checker ' +
-        'creator (arrayOf, instanceOf, objectOf, oneOf, oneOfType, and ' +
-        'shape all require an argument).',
-        componentName || 'React class',
-        ReactPropTypeLocationNames[location],
-        propName,
-        typeof error
-      );
-      if (error instanceof Error && !(error.message in loggedTypeFailures)) {
-        // Only monitor this failure once because there tends to be a lot of the
-        // same error.
-        loggedTypeFailures[error.message] = true;
-
-        warning(
-          false,
-          'Failed propType: %s%s',
-          error.message,
-          ReactComponentTreeDevtool.getCurrentStackAddendum(element)
-        );
-      }
-    }
-  }
-}
-
-/**
  * Given an element, validate that its props follow the propTypes definition,
  * provided by the type.
  *
@@ -225,9 +163,10 @@ function validatePropTypes(element) {
   }
   var name = componentClass.displayName || componentClass.name;
   if (componentClass.propTypes) {
-    checkPropTypes(
+    checkTypes(
       element,
       name,
+      element.props,
       componentClass.propTypes,
       ReactPropTypeLocations.prop
     );
