+++ src/isomorphic/classic/element/__tests__/ReactElementClone-test.js
@@ -269,7 +269,7 @@ describe('ReactElementClone', function() {
     ReactTestUtils.renderIntoDocument(React.createElement(GrandParent));
     expect(console.error.argsForCall.length).toBe(1);
     expect(console.error.argsForCall[0][0]).toBe(
-      'Warning: Failed propType: ' +
+      'Warning: Failed prop type: ' +
       'Invalid prop `color` of type `number` supplied to `Component`, ' +
       'expected `string`.\n' +
       '    in Component (created by GrandParent)\n' +
