+++ src/isomorphic/classic/element/__tests__/ReactElementValidator-test.js
@@ -281,7 +281,7 @@ describe('ReactElementValidator', function() {
     });
     ReactTestUtils.renderIntoDocument(React.createElement(ParentComp));
     expect(console.error.argsForCall[0][0]).toBe(
-      'Warning: Failed propType: ' +
+      'Warning: Failed prop type: ' +
       'Invalid prop `color` of type `number` supplied to `MyComp`, ' +
       'expected `string`.\n' +
       '    in MyComp (created by ParentComp)\n' +
@@ -360,7 +360,7 @@ describe('ReactElementValidator', function() {
 
     expect(console.error.calls.length).toBe(1);
     expect(console.error.argsForCall[0][0]).toBe(
-      'Warning: Failed propType: ' +
+      'Warning: Failed prop type: ' +
       'Required prop `prop` was not specified in `Component`.\n' +
       '    in Component'
     );
@@ -385,7 +385,7 @@ describe('ReactElementValidator', function() {
 
     expect(console.error.calls.length).toBe(1);
     expect(console.error.argsForCall[0][0]).toBe(
-      'Warning: Failed propType: ' +
+      'Warning: Failed prop type: ' +
       'Required prop `prop` was not specified in `Component`.\n' +
       '    in Component'
     );
@@ -412,13 +412,13 @@ describe('ReactElementValidator', function() {
 
     expect(console.error.calls.length).toBe(2);
     expect(console.error.argsForCall[0][0]).toBe(
-      'Warning: Failed propType: ' +
+      'Warning: Failed prop type: ' +
       'Required prop `prop` was not specified in `Component`.\n' +
       '    in Component'
     );
 
     expect(console.error.argsForCall[1][0]).toBe(
-      'Warning: Failed propType: ' +
+      'Warning: Failed prop type: ' +
       'Invalid prop `prop` of type `number` supplied to ' +
       '`Component`, expected `string`.\n' +
       '    in Component'
