+++ src/renderers/shared/stack/reconciler/ReactCompositeComponent.js
@@ -12,7 +12,6 @@
 'use strict';
 
 var ReactComponentEnvironment = require('ReactComponentEnvironment');
-var ReactComponentTreeDevtool = require('ReactComponentTreeDevtool');
 var ReactCurrentOwner = require('ReactCurrentOwner');
 var ReactElement = require('ReactElement');
 var ReactErrorUtils = require('ReactErrorUtils');
@@ -20,9 +19,10 @@ var ReactInstanceMap = require('ReactInstanceMap');
 var ReactInstrumentation = require('ReactInstrumentation');
 var ReactNodeTypes = require('ReactNodeTypes');
 var ReactPropTypeLocations = require('ReactPropTypeLocations');
-var ReactPropTypeLocationNames = require('ReactPropTypeLocationNames');
 var ReactReconciler = require('ReactReconciler');
 var ReactUpdateQueue = require('ReactUpdateQueue');
+
+var checkTypes = require('checkTypes');
 
 var emptyObject = require('emptyObject');
 var invariant = require('invariant');
@@ -665,38 +665,7 @@ var ReactCompositeComponentMixin = {
    * @private
    */
   _checkContextTypes: function(propTypes, props, location) {
-    var componentName = this.getName();
-    for (var propName in propTypes) {
-      if (propTypes.hasOwnProperty(propName)) {
-        var error;
-        try {
-          // This is intentionally an invariant that gets caught. It's the same
-          // behavior as without this statement except with a better message.
-          invariant(
-            typeof propTypes[propName] === 'function',
-            '%s: %s type `%s` is invalid; it must be a function, usually ' +
-            'from React.PropTypes.',
-            componentName || 'React class',
-            ReactPropTypeLocationNames[location],
-            propName
-          );
-          error = propTypes[propName](props, propName, componentName, location);
-        } catch (ex) {
-          error = ex;
-        }
-        if (error instanceof Error) {
-          // We may want to extend this logic for similar errors in
-          // top-level render calls, so I'm abstracting it away into
-          // a function to minimize refactoring in the future
-          warning(
-            false,
-            'Failed Context Types: %s%s',
-            error.message,
-            ReactComponentTreeDevtool.getStackAddendumByID(this._debugID)
-          );
-        }
-      }
-    }
+    checkTypes(this, this.getName(), props, propTypes, location);
   },
 
   receiveComponent: function(nextElement, transaction, nextContext) {
